requirements.txt
# Core dependencies
numpy==1.26.4
pandas==2.1.4
scipy==1.11.3

# PDF processing
pdfplumber==0.9.0
PyMuPDF==1.21.1
camelot-py==0.9.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.0.1

# ML/AI
transformers==4.35.2
sentence-transformers==2.7.0
scikit-learn==1.3.0
modAL

# PaddleOCR
paddlepaddle==2.6.1
paddleocr>=2.8.1

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.0

# Export modules
reportlab==4.0.7
python-pptx==0.6.23
openpyxl==3.1.2

# Database & Cache
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
locust==2.20.0

# Utilities
jinja2==3.1.2
ruamel.yaml==0.17.32
joblib==1.3.2
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Build tools
pyinstaller==6.2.0
eel
requests>=2.31.0
asyncpg>=0.29.0
prometheus-fastapi-instrumentator>=5.4.0
